renaming + migrate some utils from rules to util folder
